fixes #2312 : Fixed issue with Tuto15
--HG--
branch : fw4spl_0.9.1
extra : rebase_source : 7b1ba4615ed63bda679fc5f1560906ff2b5c6a46

+++ Apps/Tuto15MultithreadCtrl/src/Tuto15MultithreadCtrl/SReadArray.cpp
@@ -46,7 +46,7 @@ void SReadArray::updating() throw( ::fwTools::Failed )
 
     ::fwData::Array::SizeType size(1, arraySize);
 
-    array->resize("uint8", size, 1, true);
+    array->resize("uint32", size, 1, true);
 
     ::fwComEd::helper::Array arrayHelper(array);
